requirements.txt
SQLAlchemy
Scrapy
BeautifulSoup4
alembic
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib
